This update includes several corrections suggested by the community. Thank you everybody!
-Implementation of the getLeader command in the vehicle namespace, thanks to Marcus Müller

-Default port changed from 8813 to 8873 in traci.init, according to SUMO's latest versions

-Socket-related Java exceptions are now shown in the command window, allowing to track connection issues

-The Storage class now returns doubles instead of 32 bit integers, allowing compatibility with simulink, thanks to Pablo Deossa

-Corrected some errors regarding validation and documentation in traci.vehicle.add. depart-time is given in seconds instead of miliseconds. Thanks to the user amatsuki@hotmail.com

+++ trunk/TraCI4Matlab/+traci/Storage.m
@@ -1,7 +1,7 @@
 classdef Storage < handle
 %Storage A container for responses from SUMO server.
 
-%   Copyright 2013 Universidad Nacional de Colombia,
+%   Copyright 2015 Universidad Nacional de Colombia,
 %   Politecnico Jaime Isaza Cadavid.
 %   Authors: Andres Acosta, Jairo Espinosa, Jorge Espinosa.
 %   $Id$
@@ -28,7 +28,7 @@ classdef Storage < handle
 % 			end
         end
         function value = readInt(this)
-            value = typecast(fliplr(uint8(this.read(4))),'int32');
+            value = double(typecast(fliplr(uint8(this.read(4))),'int32'));
         end
         
         function value = readDouble(this)
